added ProxyConnectorBase
